Refactor declarative in multiple smaller modules
