+++ vhdl_lang/src/analysis/declarative.rs
@@ -6,13 +6,13 @@
 
 use super::names::*;
 use super::*;
+use crate::analysis::types::find_full_type_definition;
 use crate::ast::*;
 use crate::data::*;
 use crate::named_entity::{Signature, *};
 use crate::{ast, named_entity, HasTokenSpan};
 use analyze::*;
 use fnv::FnvHashMap;
-use itertools::Itertools;
 use std::collections::hash_map::Entry;
 
 impl Declaration {
@@ -603,251 +603,6 @@ impl<'a> AnalyzeContext<'a> {
         Ok(())
     }
 
-    /// Analyze a generic subprogram instance, i.e.,
-    /// ```vhdl
-    /// procedure my_proc is new my_proc generic map (T => std_logic);
-    /// ```
-    ///
-    /// # Arguments
-    ///
-    /// * `scope` - The scope that this instance was declared in
-    /// * `inst_subprogram_ent` - A reference to the instantiated subprogram entity.
-    ///     Used to set the parent reference of the signature
-    /// * `uninst_name` - The [ResolvedName] of the uninstantiated subprogram
-    /// * `instance` - A reference to the AST element of the subprogram instantiation
-    /// * `diagnostics` - The diagnostics handler
-    ///
-    /// # Returns
-    /// The signature after applying the optional map aspect of the uninstantiated subprogram
-    fn generic_subprogram_instance(
-        &self,
-        scope: &Scope<'a>,
-        inst_subprogram_ent: &EntRef<'a>,
-        uninst_name: &ResolvedName<'a>,
-        instance: &mut SubprogramInstantiation,
-        diagnostics: &mut dyn DiagnosticHandler,
-    ) -> AnalysisResult<Signature> {
-        let uninstantiated_subprogram =
-            self.resolve_uninstantiated_subprogram(scope, uninst_name, instance)?;
-        self.check_instantiated_subprogram_kind_matches_declared(
-            &uninstantiated_subprogram,
-            instance,
-            diagnostics,
-        );
-        instance
-            .subprogram_name
-            .item
-            .set_unique_reference(&uninstantiated_subprogram);
-        let region = match uninstantiated_subprogram.kind() {
-            Overloaded::UninstSubprogramDecl(_, region) => region,
-            Overloaded::UninstSubprogram(_, region) => region,
-            _ => unreachable!(),
-        };
-
-        match as_fatal(self.generic_instance(
-            inst_subprogram_ent,
-            scope,
-            &instance.ident.tree.pos,
-            region,
-            &mut instance.generic_map,
-            diagnostics,
-        ))? {
-            None => Ok(uninstantiated_subprogram.signature().clone()),
-            Some((_, mapping)) => {
-                match self.map_signature(
-                    Some(inst_subprogram_ent),
-                    &mapping,
-                    uninstantiated_subprogram.signature(),
-                ) {
-                    Ok(signature) => Ok(signature),
-                    Err(err) => {
-                        let mut diag = Diagnostic::error(&instance.ident.tree.pos, err);
-                        if let Some(pos) = uninstantiated_subprogram.decl_pos() {
-                            diag.add_related(pos, "When instantiating this declaration");
-                        }
-                        Err(AnalysisError::NotFatal(diag))
-                    }
-                }
-            }
-        }
-    }
-
-    /// Given a `ResolvedName` and the subprogram instantiation,
-    /// find the uninstantiated subprogram that the resolved name references.
-    /// Return that resolved subprogram, if it exists, else return an `Err`
-    fn resolve_uninstantiated_subprogram(
-        &self,
-        scope: &Scope<'a>,
-        name: &ResolvedName<'a>,
-        instantiation: &mut SubprogramInstantiation,
-    ) -> AnalysisResult<OverloadedEnt<'a>> {
-        let signature_key = match &mut instantiation.signature {
-            None => None,
-            Some(ref mut signature) => Some((
-                self.resolve_signature(scope, signature)?,
-                signature.pos.clone(),
-            )),
-        };
-        let overloaded_ent = match name {
-            ResolvedName::Overloaded(_, overloaded) => {
-                let choices = overloaded
-                    .entities()
-                    .filter(|ent| ent.is_uninst_subprogram())
-                    .collect_vec();
-                if choices.is_empty() {
-                    Err(AnalysisError::NotFatal(Diagnostic::error(
-                        &instantiation.ident.tree.pos,
-                        format!(
-                            "{} does not denote an uninstantiated subprogram",
-                            name.describe()
-                        ),
-                    )))
-                } else if choices.len() == 1 {
-                    // There is only one possible candidate
-                    let ent = choices[0];
-                    // If the instantiated program has a signature, check that it matches
-                    // that of the uninstantiated subprogram
-                    if let Some((key, pos)) = signature_key {
-                        match overloaded.get(&SubprogramKey::Uninstantiated(key)) {
-                            None => Err(AnalysisError::NotFatal(Diagnostic::error(
-                                pos.clone(),
-                                format!(
-                                    "Signature does not match the the signature of {}",
-                                    ent.describe()
-                                ),
-                            ))),
-                            Some(_) => Ok(ent),
-                        }
-                    } else {
-                        Ok(ent)
-                    }
-                } else if let Some((key, _)) = signature_key {
-                    // There are multiple candidates
-                    // but there is a signature that we can try to resolve
-                    if let Some(resolved_ent) =
-                        overloaded.get(&SubprogramKey::Uninstantiated(key.clone()))
-                    {
-                        Ok(resolved_ent)
-                    } else {
-                        Err(AnalysisError::NotFatal(Diagnostic::error(
-                            &instantiation.subprogram_name.pos,
-                            format!(
-                                "No uninstantiated subprogram exists with signature {}",
-                                key.describe()
-                            ),
-                        )))
-                    }
-                } else {
-                    // There are multiple candidates
-                    // and there is no signature to resolve
-                    let mut err = Diagnostic::error(
-                        &instantiation.subprogram_name.pos,
-                        format!("Ambiguous instantiation of '{}'", overloaded.designator()),
-                    );
-                    for ent in choices {
-                        if let Some(pos) = &ent.decl_pos {
-                            err.add_related(pos.clone(), format!("Might be {}", ent.describe()))
-                        }
-                    }
-                    Err(AnalysisError::NotFatal(err))
-                }
-            }
-            _ => Err(AnalysisError::NotFatal(Diagnostic::error(
-                &instantiation.subprogram_name.pos,
-                format!(
-                    "{} does not denote an uninstantiated subprogram",
-                    name.describe()
-                ),
-            ))),
-        }?;
-        if overloaded_ent.is_uninst_subprogram() {
-            Ok(overloaded_ent)
-        } else {
-            Err(AnalysisError::NotFatal(Diagnostic::error(
-                &instantiation.subprogram_name.pos,
-                format!("{} cannot be instantiated", overloaded_ent.describe()),
-            )))
-        }
-    }
-
-    /// Checks that an instantiated subprogram kind matches the declared subprogram.
-    /// For instance, when a subprogram was instantiated using
-    /// ```vhdl
-    /// function my_func is new proc;
-    /// ```
-    /// where proc is
-    /// ```vhdl
-    /// procedure proc is
-    /// ...
-    /// ```
-    ///
-    /// This function will push an appropriate diagnostic.
-    fn check_instantiated_subprogram_kind_matches_declared(
-        &self,
-        ent: &OverloadedEnt,
-        instance: &SubprogramInstantiation,
-        diagnostics: &mut dyn DiagnosticHandler,
-    ) {
-        let err_msg = if ent.is_function() && instance.kind != SubprogramKind::Function {
-            Some("Instantiating function as procedure")
-        } else if ent.is_procedure() && instance.kind != SubprogramKind::Procedure {
-            Some("Instantiating procedure as function")
-        } else {
-            None
-        };
-        if let Some(msg) = err_msg {
-            let mut err = Diagnostic::error(self.ctx.get_pos(instance.get_start_token()), msg);
-            if let Some(pos) = ent.decl_pos() {
-                err.add_related(pos, format!("{} declared here", ent.describe()));
-            }
-            diagnostics.push(err)
-        }
-    }
-
-    fn find_subpgm_specification(
-        &self,
-        scope: &Scope<'a>,
-        decl: &SubprogramSpecification,
-        signature: &Signature,
-    ) -> Option<OverloadedEnt<'a>> {
-        let des = decl.subpgm_designator().item.clone().into_designator();
-
-        if let Some(NamedEntities::Overloaded(overloaded)) = scope.lookup_immediate(&des) {
-            let ent = overloaded.get(&SubprogramKey::Normal(signature.key()))?;
-
-            if ent.is_subprogram_decl() {
-                return Some(ent);
-            }
-        }
-        None
-    }
-
-    fn find_uninst_subpgm_specification(
-        &self,
-        scope: &Scope<'a>,
-        decl: &SubprogramSpecification,
-        signature: &Signature,
-    ) -> Option<OverloadedEnt<'a>> {
-        let des = decl.subpgm_designator().item.clone().into_designator();
-
-        if let Some(NamedEntities::Overloaded(overloaded)) = scope.lookup_immediate(&des) {
-            // Note: This does not work in common circumstances with a generic type parameter
-            // since the parameters of the declared subprogram and the subprogram with body
-            // point to two different type-ID's. For example:
-            // function foo generic (type F);
-            //                            ^-- F has EntityId X
-            // function foo generic (type F) return F is ... end function foo;
-            //                            ^-- F has EntityId Y
-            // A future improvement must take this fact into account.
-            let ent = overloaded.get(&SubprogramKey::Uninstantiated(signature.key()))?;
-
-            if ent.is_uninst_subprogram_decl() {
-                return Some(ent);
-            }
-        }
-        None
-    }
-
     fn find_deferred_constant_declaration(
         &self,
         scope: &Scope<'a>,
@@ -1022,509 +777,7 @@ impl<'a> AnalyzeContext<'a> {
         Ok(())
     }
 
-    pub(crate) fn analyze_type_declaration(
-        &self,
-        scope: &Scope<'a>,
-        parent: EntRef<'a>,
-        type_decl: &mut TypeDeclaration,
-        // Is the full type declaration of an incomplete type
-        // Overwrite id when defining full type
-        overwrite_id: Option<EntityId>,
-        diagnostics: &mut dyn DiagnosticHandler,
-    ) -> FatalResult {
-        match type_decl.def {
-            TypeDefinition::Enumeration(ref mut enumeration) => {
-                let enum_type = TypeEnt::define_with_opt_id(
-                    self.arena,
-                    overwrite_id,
-                    &mut type_decl.ident,
-                    parent,
-                    None,
-                    Type::Enum(
-                        enumeration
-                            .iter()
-                            .map(|literal| literal.tree.item.clone().into_designator())
-                            .collect(),
-                    ),
-                );
-
-                let signature =
-                    Signature::new(FormalRegion::new(InterfaceType::Parameter), Some(enum_type));
-
-                for literal in enumeration.iter_mut() {
-                    let literal_ent = self.arena.explicit(
-                        literal.tree.item.clone().into_designator(),
-                        enum_type.into(),
-                        AnyEntKind::Overloaded(Overloaded::EnumLiteral(signature.clone())),
-                        Some(&literal.tree.pos),
-                    );
-                    literal.decl.set(literal_ent.id());
-
-                    unsafe {
-                        self.arena.add_implicit(enum_type.id(), literal_ent);
-                    }
-
-                    scope.add(literal_ent, diagnostics);
-                }
-
-                scope.add(enum_type.into(), diagnostics);
-
-                for ent in self.enum_implicits(enum_type, self.has_matching_op(enum_type)) {
-                    unsafe {
-                        self.arena.add_implicit(enum_type.id(), ent);
-                    }
-
-                    scope.add(ent, diagnostics);
-                }
-            }
-            TypeDefinition::ProtectedBody(ref mut body) => {
-                match scope.lookup_immediate(&type_decl.ident.tree.item.clone().into()) {
-                    Some(visible) => {
-                        let is_ok = match visible.clone().into_non_overloaded() {
-                            Ok(ent) => {
-                                if let AnyEntKind::Type(Type::Protected(ptype_region, is_body)) =
-                                    ent.kind()
-                                {
-                                    if *is_body {
-                                        if let Some(prev_pos) = ent.decl_pos() {
-                                            diagnostics.push(Diagnostic::duplicate_error(
-                                                &type_decl.ident.tree,
-                                                &type_decl.ident.tree.pos,
-                                                Some(prev_pos),
-                                            ))
-                                        }
-                                    } else {
-                                        let ptype_body: &'a AnyEnt = TypeEnt::define_with_opt_id(
-                                            self.arena,
-                                            overwrite_id,
-                                            &mut type_decl.ident,
-                                            parent,
-                                            Some(ent),
-                                            Type::Protected(Region::default(), true),
-                                        )
-                                        .into();
-
-                                        let region = Scope::extend(ptype_region, Some(scope));
-                                        self.analyze_declarative_part(
-                                            &region,
-                                            ptype_body,
-                                            &mut body.decl,
-                                            diagnostics,
-                                        )?;
-
-                                        let kind = Type::Protected(region.into_region(), true);
-                                        unsafe {
-                                            ptype_body.set_kind(AnyEntKind::Type(kind));
-                                        }
-
-                                        scope.add(ptype_body, diagnostics);
-                                    }
-
-                                    true
-                                } else {
-                                    false
-                                }
-                            }
-                            _ => false,
-                        };
-
-                        if !is_ok {
-                            diagnostics.push(Diagnostic::error(
-                                type_decl.ident.pos(),
-                                format!("'{}' is not a protected type", &type_decl.ident),
-                            ));
-                        }
-                    }
-                    None => {
-                        diagnostics.push(Diagnostic::error(
-                            type_decl.ident.pos(),
-                            format!("No declaration of protected type '{}'", &type_decl.ident),
-                        ));
-                    }
-                };
-            }
-            TypeDefinition::Protected(ref mut prot_decl) => {
-                // Protected type name is visible inside its declarative region
-                // This will be overwritten later when the protected type region is finished
-                let ptype: &'a AnyEnt = TypeEnt::define_with_opt_id(
-                    self.arena,
-                    overwrite_id,
-                    &mut type_decl.ident,
-                    parent,
-                    None,
-                    Type::Protected(Region::default(), false),
-                )
-                .into();
-
-                scope.add(ptype, diagnostics);
-
-                let region = scope.nested();
-                for item in prot_decl.items.iter_mut() {
-                    match item {
-                        ProtectedTypeDeclarativeItem::Subprogram(ref mut subprogram) => {
-                            match self.subprogram_specification(
-                                scope,
-                                ptype,
-                                &mut subprogram.specification,
-                                Overloaded::SubprogramDecl,
-                                diagnostics,
-                            ) {
-                                Ok((_, ent)) => {
-                                    region.add(ent.into(), diagnostics);
-                                }
-                                Err(err) => {
-                                    diagnostics.push(err.into_non_fatal()?);
-                                    return Ok(());
-                                }
-                            }
-                        }
-                    }
-                }
-
-                // This is safe since we are in a single thread and no other reference can exist yes
-                // Also the region is stored inside an Arc which cannot move
-                {
-                    let AnyEntKind::Type(Type::Protected(region_ptr, _)) = ptype.kind() else {
-                        unreachable!();
-                    };
-
-                    let region_ptr = unsafe {
-                        let region_ptr = region_ptr as *const Region;
-                        let region_ptr = region_ptr as *mut Region;
-                        &mut *region_ptr as &mut Region
-                    };
-                    *region_ptr = region.into_region();
-                }
-            }
-            TypeDefinition::Record(ref mut element_decls) => {
-                let type_ent = TypeEnt::define_with_opt_id(
-                    self.arena,
-                    overwrite_id,
-                    &mut type_decl.ident,
-                    parent,
-                    None,
-                    Type::Record(RecordRegion::default()),
-                );
-
-                let mut elems = RecordRegion::default();
-                let mut region = Region::default();
-                for elem_decl in element_decls.iter_mut() {
-                    let subtype =
-                        self.resolve_subtype_indication(scope, &mut elem_decl.subtype, diagnostics);
-                    match subtype {
-                        Ok(subtype) => {
-                            let elem = self.arena.define(
-                                &mut elem_decl.ident,
-                                type_ent.into(),
-                                AnyEntKind::ElementDeclaration(subtype),
-                            );
-                            region.add(elem, diagnostics);
-                            elems.add(elem);
-                        }
-                        Err(err) => {
-                            err.add_to(diagnostics)?;
-                        }
-                    }
-                }
-                region.close(diagnostics);
-
-                unsafe {
-                    let kind = AnyEntKind::Type(Type::Record(elems));
-                    type_ent.set_kind(kind)
-                }
-
-                scope.add(type_ent.into(), diagnostics);
-
-                for ent in self.record_implicits(type_ent) {
-                    unsafe {
-                        self.arena.add_implicit(type_ent.id(), ent);
-                    }
-                    scope.add(ent, diagnostics);
-                }
-            }
-            TypeDefinition::Access(ref mut subtype_indication) => {
-                let subtype =
-                    self.resolve_subtype_indication(scope, subtype_indication, diagnostics);
-                match subtype {
-                    Ok(subtype) => {
-                        let type_ent = TypeEnt::define_with_opt_id(
-                            self.arena,
-                            overwrite_id,
-                            &mut type_decl.ident,
-                            parent,
-                            None,
-                            Type::Access(subtype),
-                        );
-
-                        scope.add(type_ent.into(), diagnostics);
-
-                        for ent in self.access_implicits(type_ent) {
-                            unsafe {
-                                self.arena.add_implicit(type_ent.id(), ent);
-                            }
-                            scope.add(ent, diagnostics);
-                        }
-                    }
-                    Err(err) => err.add_to(diagnostics)?,
-                }
-            }
-            TypeDefinition::Array(ref mut array_indexes, ref mut subtype_indication) => {
-                let mut indexes: Vec<Option<BaseType>> = Vec::with_capacity(array_indexes.len());
-                for index in array_indexes.iter_mut() {
-                    indexes.push(as_fatal(self.analyze_array_index(
-                        scope,
-                        index,
-                        diagnostics,
-                    ))?);
-                }
-
-                let elem_type =
-                    match self.resolve_subtype_indication(scope, subtype_indication, diagnostics) {
-                        Ok(subtype) => subtype.type_mark().to_owned(),
-                        Err(err) => {
-                            err.add_to(diagnostics)?;
-                            return Ok(());
-                        }
-                    };
-
-                let is_1d = indexes.len() == 1;
-                let array_ent = TypeEnt::define_with_opt_id(
-                    self.arena,
-                    overwrite_id,
-                    &mut type_decl.ident,
-                    parent,
-                    None,
-                    Type::Array { indexes, elem_type },
-                );
-
-                scope.add(array_ent.into(), diagnostics);
-
-                for ent in self.array_implicits(array_ent, is_1d && self.has_matching_op(elem_type))
-                {
-                    unsafe {
-                        self.arena.add_implicit(array_ent.id(), ent);
-                    }
-                    scope.add(ent, diagnostics);
-                }
-            }
-            TypeDefinition::Subtype(ref mut subtype_indication) => {
-                match self.resolve_subtype_indication(scope, subtype_indication, diagnostics) {
-                    Ok(subtype) => {
-                        let type_ent = TypeEnt::define_with_opt_id(
-                            self.arena,
-                            overwrite_id,
-                            &mut type_decl.ident,
-                            parent,
-                            None,
-                            Type::Subtype(subtype),
-                        );
-                        scope.add(type_ent.into(), diagnostics);
-                    }
-                    Err(err) => {
-                        err.add_to(diagnostics)?;
-                    }
-                }
-            }
-            TypeDefinition::Physical(ref mut physical) => {
-                self.range_with_ttyp(
-                    scope,
-                    self.universal_integer().into(),
-                    &mut physical.range,
-                    diagnostics,
-                )?;
-
-                let phys_type = TypeEnt::define_with_opt_id(
-                    self.arena,
-                    overwrite_id,
-                    &mut type_decl.ident,
-                    parent,
-                    None,
-                    Type::Physical,
-                );
-                scope.add(phys_type.into(), diagnostics);
-
-                let primary = self.arena.define(
-                    &mut physical.primary_unit,
-                    parent,
-                    AnyEntKind::PhysicalLiteral(phys_type),
-                );
-
-                unsafe {
-                    self.arena.add_implicit(phys_type.id(), primary);
-                }
-                scope.add(primary, diagnostics);
-
-                for (secondary_unit_name, value) in physical.secondary_units.iter_mut() {
-                    match self.resolve_physical_unit(scope, &mut value.unit) {
-                        Ok(secondary_unit_type) => {
-                            if secondary_unit_type.base_type() != phys_type {
-                                diagnostics.error(
-                                    &value.unit.item.pos,
-                                    format!(
-                                        "Physical unit of type '{}' does not match {}",
-                                        secondary_unit_type.designator(),
-                                        phys_type.describe()
-                                    ),
-                                )
-                            }
-                        }
-                        Err(err) => diagnostics.push(err),
-                    }
-
-                    let secondary_unit = self.arena.define(
-                        secondary_unit_name,
-                        parent,
-                        AnyEntKind::PhysicalLiteral(phys_type),
-                    );
-                    unsafe {
-                        self.arena.add_implicit(phys_type.id(), secondary_unit);
-                    }
-                    scope.add(secondary_unit, diagnostics)
-                }
-
-                for ent in self.physical_implicits(phys_type) {
-                    unsafe {
-                        self.arena.add_implicit(phys_type.id(), ent);
-                    }
-                    scope.add(ent, diagnostics);
-                }
-            }
-            TypeDefinition::Incomplete(..) => {
-                unreachable!("Handled elsewhere");
-            }
-
-            TypeDefinition::Numeric(ref mut range) => {
-                self.range_unknown_typ(scope, range, diagnostics)?;
-
-                let universal_type = if let Some(range_typ) =
-                    as_fatal(self.range_type(scope, range, diagnostics))?
-                {
-                    if range_typ.is_any_integer() {
-                        UniversalType::Integer
-                    } else if range_typ.is_any_real() {
-                        UniversalType::Real
-                    } else {
-                        diagnostics.error(&range.pos(), "Expected real or integer range");
-                        return Ok(());
-                    }
-                } else {
-                    return Ok(());
-                };
-
-                let type_ent = TypeEnt::define_with_opt_id(
-                    self.arena,
-                    overwrite_id,
-                    &mut type_decl.ident,
-                    parent,
-                    None,
-                    match universal_type {
-                        UniversalType::Integer => Type::Integer,
-                        UniversalType::Real => Type::Real,
-                    },
-                );
-                scope.add(type_ent.into(), diagnostics);
-
-                for ent in self.numeric_implicits(universal_type, type_ent) {
-                    unsafe {
-                        self.arena.add_implicit(type_ent.id(), ent);
-                    }
-                    scope.add(ent, diagnostics);
-                }
-            }
-
-            TypeDefinition::File(ref mut type_mark) => {
-                let file_type = TypeEnt::define_with_opt_id(
-                    self.arena,
-                    overwrite_id,
-                    &mut type_decl.ident,
-                    parent,
-                    None,
-                    Type::File,
-                );
-
-                match self.resolve_type_mark(scope, type_mark) {
-                    Ok(type_mark) => {
-                        for ent in self.create_implicit_file_type_subprograms(file_type, type_mark)
-                        {
-                            unsafe {
-                                self.arena.add_implicit(file_type.id(), ent);
-                            }
-                            scope.add(ent, diagnostics);
-                        }
-                    }
-                    Err(err) => {
-                        err.add_to(diagnostics)?;
-                    }
-                }
-
-                scope.add(file_type.into(), diagnostics);
-            }
-        }
-
-        Ok(())
-    }
-
-    /// The matching operators such as ?= are defined for 1d arrays of bit and std_ulogic element type
-    fn has_matching_op(&self, typ: TypeEnt<'a>) -> bool {
-        if self.is_std_logic_1164 {
-            // Within the std_logic_1164 we do not have efficient access to the types
-            typ.designator() == &Designator::Identifier(self.root.symbol_utf8("std_ulogic"))
-        } else {
-            if let Some(ref standard_types) = self.root.standard_types {
-                if typ.id() == standard_types.bit {
-                    return true;
-                }
-            }
-
-            if let Some(id) = self.root.std_ulogic {
-                if typ.id() == id {
-                    return true;
-                }
-            }
-
-            false
-        }
-    }
-
-    pub fn resolve_signature(
-        &self,
-        scope: &Scope<'a>,
-        signature: &mut WithPos<ast::Signature>,
-    ) -> AnalysisResult<SignatureKey> {
-        let (args, return_type) = match &mut signature.item {
-            ast::Signature::Function(ref mut args, ref mut ret) => {
-                let args: Vec<_> = args
-                    .iter_mut()
-                    .map(|arg| self.resolve_type_mark(scope, arg))
-                    .collect();
-                let return_type = self.resolve_type_mark(scope, ret);
-                (args, Some(return_type))
-            }
-            ast::Signature::Procedure(args) => {
-                let args: Vec<_> = args
-                    .iter_mut()
-                    .map(|arg| self.resolve_type_mark(scope, arg))
-                    .collect();
-                (args, None)
-            }
-        };
-
-        let mut params = Vec::with_capacity(args.len());
-        for arg in args {
-            params.push(arg?.base());
-        }
-
-        if let Some(return_type) = return_type {
-            Ok(SignatureKey::new(
-                params,
-                Some(return_type?.base_type().base()),
-            ))
-        } else {
-            Ok(SignatureKey::new(params, None))
-        }
-    }
-
-    fn analyze_interface_declaration(
+    pub fn analyze_interface_declaration(
         &self,
         scope: &Scope<'a>,
         parent: EntRef<'a>,
@@ -1670,7 +923,7 @@ impl<'a> AnalyzeContext<'a> {
         Ok(params)
     }
 
-    fn analyze_array_index(
+    pub(crate) fn analyze_array_index(
         &self,
         scope: &Scope<'a>,
         array_index: &mut ArrayIndex,
@@ -1689,295 +942,6 @@ impl<'a> AnalyzeContext<'a> {
             ArrayIndex::Discrete(ref mut drange) => self.drange_type(scope, drange, diagnostics),
         }
     }
-
-    fn analyze_subtype_constraint(
-        &self,
-        scope: &Scope<'a>,
-        pos: &SrcPos, // The position of the root type mark
-        base_type: BaseType<'a>,
-        constraint: &mut SubtypeConstraint,
-        diagnostics: &mut dyn DiagnosticHandler,
-    ) -> FatalResult {
-        match constraint {
-            SubtypeConstraint::Array(ref mut dranges, ref mut constraint) => {
-                if let Type::Array { indexes, elem_type } = base_type.kind() {
-                    for (idx, drange) in dranges.iter_mut().enumerate() {
-                        if let Some(index_typ) = indexes.get(idx) {
-                            if let Some(index_typ) = index_typ {
-                                self.drange_with_ttyp(
-                                    scope,
-                                    (*index_typ).into(),
-                                    drange,
-                                    diagnostics,
-                                )?;
-                            } else {
-                                self.drange_unknown_type(scope, drange, diagnostics)?;
-                            }
-                        } else {
-                            diagnostics.error(
-                                drange.pos(),
-                                format!("Got extra index constraint for {}", base_type.describe()),
-                            );
-                        }
-                    }
-
-                    // empty dranges means (open)
-                    if dranges.len() < indexes.len() && !dranges.is_empty() {
-                        diagnostics.error(
-                            pos,
-                            format!(
-                                "Too few index constraints for {}. Got {} but expected {}",
-                                base_type.describe(),
-                                dranges.len(),
-                                indexes.len()
-                            ),
-                        );
-                    }
-
-                    if let Some(constraint) = constraint {
-                        self.analyze_subtype_constraint(
-                            scope,
-                            &constraint.pos,
-                            elem_type.base(),
-                            &mut constraint.item,
-                            diagnostics,
-                        )?;
-                    }
-                } else {
-                    diagnostics.error(
-                        pos,
-                        format!(
-                            "Array constraint cannot be used for {}",
-                            base_type.describe()
-                        ),
-                    );
-                }
-            }
-            SubtypeConstraint::Range(ref mut range) => {
-                if base_type.is_scalar() {
-                    self.range_with_ttyp(scope, base_type.into(), range, diagnostics)?;
-                } else {
-                    diagnostics.error(
-                        pos,
-                        format!(
-                            "Scalar constraint cannot be used for {}",
-                            base_type.describe()
-                        ),
-                    );
-                }
-            }
-            SubtypeConstraint::Record(ref mut constraints) => {
-                if let Type::Record(region) = base_type.kind() {
-                    for constraint in constraints.iter_mut() {
-                        let ElementConstraint { ident, constraint } = constraint;
-                        let des = Designator::Identifier(ident.item.clone());
-                        if let Some(elem) = region.lookup(&des) {
-                            self.analyze_subtype_constraint(
-                                scope,
-                                &constraint.pos,
-                                elem.type_mark().base(),
-                                &mut constraint.item,
-                                diagnostics,
-                            )?;
-                        } else {
-                            diagnostics.push(Diagnostic::no_declaration_within(
-                                &base_type, &ident.pos, &des,
-                            ))
-                        }
-                    }
-                } else {
-                    diagnostics.error(
-                        pos,
-                        format!(
-                            "Record constraint cannot be used for {}",
-                            base_type.describe()
-                        ),
-                    );
-                }
-            }
-        }
-        Ok(())
-    }
-
-    pub fn resolve_subtype_indication(
-        &self,
-        scope: &Scope<'a>,
-        subtype_indication: &mut SubtypeIndication,
-        diagnostics: &mut dyn DiagnosticHandler,
-    ) -> AnalysisResult<Subtype<'a>> {
-        // @TODO more
-        let SubtypeIndication {
-            type_mark,
-            constraint,
-            ..
-        } = subtype_indication;
-
-        let base_type = self.resolve_type_mark(scope, type_mark)?;
-
-        if let Some(constraint) = constraint {
-            self.analyze_subtype_constraint(
-                scope,
-                &type_mark.pos,
-                base_type.base(),
-                &mut constraint.item,
-                diagnostics,
-            )?;
-        }
-
-        Ok(Subtype::new(base_type))
-    }
-
-    pub fn analyze_subtype_indication(
-        &self,
-        scope: &Scope<'a>,
-        subtype_indication: &mut SubtypeIndication,
-        diagnostics: &mut dyn DiagnosticHandler,
-    ) -> FatalResult {
-        if let Err(err) = self.resolve_subtype_indication(scope, subtype_indication, diagnostics) {
-            err.add_to(diagnostics)?;
-        }
-        Ok(())
-    }
-
-    fn subprogram_header(
-        &self,
-        scope: &Scope<'a>,
-        parent: EntRef<'a>,
-        header: &mut SubprogramHeader,
-        diagnostics: &mut dyn DiagnosticHandler,
-    ) -> FatalResult<Region<'a>> {
-        let mut region = Region::default();
-        for decl in header.generic_list.iter_mut() {
-            match self.analyze_interface_declaration(scope, parent, decl, diagnostics) {
-                Ok(ent) => {
-                    region.add(ent, diagnostics);
-                    scope.add(ent, diagnostics);
-                }
-                Err(err) => {
-                    err.add_to(diagnostics)?;
-                }
-            }
-        }
-        self.analyze_map_aspect(scope, &mut header.map_aspect, diagnostics)?;
-        Ok(region)
-    }
-
-    fn subprogram_specification(
-        &self,
-        scope: &Scope<'a>,
-        parent: EntRef<'a>,
-        subprogram: &mut SubprogramSpecification,
-        to_kind: impl Fn(Signature<'a>) -> Overloaded<'a>,
-        diagnostics: &mut dyn DiagnosticHandler,
-    ) -> AnalysisResult<(Scope<'a>, OverloadedEnt<'a>)> {
-        let subpgm_region = scope.nested();
-        let ent = self.arena.explicit(
-            subprogram
-                .subpgm_designator()
-                .item
-                .clone()
-                .into_designator(),
-            parent,
-            AnyEntKind::Overloaded(to_kind(Signature::new(FormalRegion::new_params(), None))),
-            Some(&subprogram.subpgm_designator().pos),
-        );
-
-        let (signature, generic_map) = match subprogram {
-            SubprogramSpecification::Function(fun) => {
-                let generic_map = if let Some(header) = &mut fun.header {
-                    Some(self.subprogram_header(&subpgm_region, ent, header, diagnostics)?)
-                } else {
-                    None
-                };
-                let params = self.analyze_parameter_list(
-                    &subpgm_region,
-                    ent,
-                    &mut fun.parameter_list,
-                    diagnostics,
-                );
-                let return_type = self.resolve_type_mark(&subpgm_region, &mut fun.return_type);
-                (Signature::new(params?, Some(return_type?)), generic_map)
-            }
-            SubprogramSpecification::Procedure(procedure) => {
-                let generic_map = if let Some(header) = &mut procedure.header {
-                    Some(self.subprogram_header(&subpgm_region, ent, header, diagnostics)?)
-                } else {
-                    None
-                };
-                let params = self.analyze_parameter_list(
-                    &subpgm_region,
-                    ent,
-                    &mut procedure.parameter_list,
-                    diagnostics,
-                );
-                (Signature::new(params?, None), generic_map)
-            }
-        };
-
-        let mut kind = to_kind(signature);
-        if let Some(map) = generic_map {
-            match kind {
-                Overloaded::SubprogramDecl(signature) => {
-                    kind = Overloaded::UninstSubprogramDecl(signature, map)
-                }
-                Overloaded::Subprogram(signature) => {
-                    kind = Overloaded::UninstSubprogram(signature, map)
-                }
-                _ => unreachable!(),
-            }
-        }
-
-        match kind {
-            Overloaded::Subprogram(_) => {
-                let declared_by =
-                    self.find_subpgm_specification(scope, subprogram, kind.signature());
-
-                if let Some(declared_by) = declared_by {
-                    unsafe {
-                        ent.set_declared_by(declared_by.into());
-                    }
-                }
-            }
-            Overloaded::UninstSubprogram(_, _) => {
-                let declared_by =
-                    self.find_uninst_subpgm_specification(scope, subprogram, kind.signature());
-
-                if let Some(declared_by) = declared_by {
-                    unsafe {
-                        ent.set_declared_by(declared_by.into());
-                    }
-                }
-            }
-            _ => {}
-        }
-
-        unsafe {
-            ent.set_kind(AnyEntKind::Overloaded(kind));
-        }
-        subprogram.set_decl_id(ent.id());
-        Ok((subpgm_region, OverloadedEnt::from_any(ent).unwrap()))
-    }
-}
-
-fn find_full_type_definition<'a>(
-    name: &Symbol,
-    decls: &'a [Declaration],
-) -> Option<&'a TypeDeclaration> {
-    for decl in decls.iter() {
-        if let Declaration::Type(type_decl) = decl {
-            match type_decl.def {
-                TypeDefinition::Incomplete(..) => {
-                    // ignored
-                }
-                _ => {
-                    if type_decl.ident.name() == name {
-                        return Some(type_decl);
-                    }
-                }
-            }
-        }
-    }
-    None
 }
 
 impl Diagnostic {
